[WICS-6] Formatting changes
