+++ wic/src/main/java/edu/asu/diging/wic/core/controller/DeleteConceptTextController.java
@@ -1,7 +1,5 @@
 package edu.asu.diging.wic.core.controller;
 
-import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,7 +10,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import edu.asu.diging.wic.core.model.impl.ConceptText;
 import edu.asu.diging.wic.core.service.IConceptTextService;
 
 @Controller
@@ -27,8 +24,6 @@ public class DeleteConceptTextController {
 			@RequestParam(value = "pageNumber", required = true)String pageNumber, Model model) {
 		
 		conceptTextService.deleteText(id);
-		List<ConceptText> allConceptText = conceptTextService.findAll(pageNumber);
-		model.addAttribute("allConceptText",allConceptText);
 		return new ResponseEntity<>(HttpStatus.OK);
 	}
 }
