+++ wic/src/main/java/edu/asu/diging/wic/core/controller/UpdateConceptTextController.java
@@ -6,6 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
@@ -19,17 +21,29 @@ public class UpdateConceptTextController {
 	@Autowired
 	private IConceptTextService conceptTextService;
 
+    @RequestMapping(value="/admin/text/{id}/edit", method=RequestMethod.GET)
+    public String editConceptTextView( @PathVariable("id") String id, Model model,Principal principal) {
+        
+        ConceptText dataObj = conceptTextService.getTextById(id);
+        model.addAttribute("idOfData", dataObj.getId());
+        model.addAttribute("title", dataObj.getTitle());
+        model.addAttribute("text", dataObj.getText());
+        model.addAttribute("conceptId", dataObj.getConceptId());
+        model.addAttribute("author", dataObj.getAuthor());
+        return "admin/text/edit";
+    }
+
 	@RequestMapping(value="/admin/text/update", method=RequestMethod.POST)
 	public ResponseEntity<String> updateTextData(ConceptText updatedForm,
 			Principal principal, RedirectAttributes redirectAttrs) {
 		
-		if(updatedForm.getAuthor() == null || updatedForm.getAuthor().trim().isEmpty()) {
-			return new ResponseEntity<>("Author missing",HttpStatus.NO_CONTENT);
-		}
-		if(updatedForm.getTitle() == null || updatedForm.getTitle().trim().isEmpty()) {
-			return new ResponseEntity<>("Title missing",HttpStatus.NO_CONTENT);
-		}
-		conceptTextService.updateText(updatedForm, principal.getName());
+        if(updatedForm.getAuthor() == null || updatedForm.getAuthor().trim().isEmpty()) {
+        	return new ResponseEntity<>("Author missing",HttpStatus.NO_CONTENT);
+        }
+        if(updatedForm.getTitle() == null || updatedForm.getTitle().trim().isEmpty()) {
+        	return new ResponseEntity<>("Title missing",HttpStatus.NO_CONTENT);
+        }
+        conceptTextService.updateText(updatedForm, principal.getName());
 	    return new ResponseEntity<>(HttpStatus.OK);
 	}
 }
